Move src into semantic directories
