+++ packages/airview-cms/src/app/index.js
@@ -1,0 +1,1 @@
+export { AirviewCMS } from "./airview-cms";
