+++ packages/airview-cms/src/features/airview-provider/airview-provider.js
@@ -2,14 +2,9 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Provider } from "react-redux";
 import { airviewStore } from "../airview-store";
-import { CMS } from "../cms";
 
 export function AirviewProvider({ children }) {
-  return (
-    <Provider store={airviewStore}>
-      <CMS>{children}</CMS>
-    </Provider>
-  );
+  return <Provider store={airviewStore}>{children}</Provider>;
 }
 
 AirviewProvider.propTypes = {
