+++ packages/airview-cms/src/features/cms/index.js
@@ -1,4 +1,5 @@
 export { CMS } from "./cms";
+export { EnableCmsButton } from "./enable-cms-button";
 export {
   cmsSlice,
   enableCms,
