+++ packages/airview-cms/src/index.js
@@ -1,1 +1,4 @@
-export * from "@features";
+import { AirviewCMS } from "@app";
+
+export default AirviewCMS;
+export { useGetAllEntriesMeta, useGetEntry } from "@features";
