+++ packages/airview-cms/src/views/index.js
@@ -1,0 +1,1 @@
+export * from "./main-view";
