+++ packages/airview-cms/src/views/main-view/index.js
@@ -1,0 +1,1 @@
+export { MainView } from "./main-view";
