+++ packages/airview-cms/src/views/main-view/main-view.js
@@ -1,0 +1,23 @@
+import React from "react";
+import PropTypes from "prop-types";
+import { useSelector } from "react-redux";
+import { CMS, EnableCmsButton, selectCmsEnabledStatus } from "@features";
+
+export function MainView({ children }) {
+  const cmsEnabled = useSelector(selectCmsEnabledStatus);
+
+  if (!cmsEnabled) {
+    return (
+      <React.Fragment>
+        {children}
+        <EnableCmsButton />
+      </React.Fragment>
+    );
+  }
+
+  return <CMS>{children}</CMS>;
+}
+
+MainView.propTypes = {
+  children: PropTypes.node,
+};
